+++ src/lib.rs
@@ -1,15 +1,14 @@
 pub mod commands;
 pub mod config;
 pub mod constants;
-pub mod errors;
 pub mod git;
 pub mod scanner;
 pub mod utils;
 
 use config::Config;
-use errors::AppResult;
+use eyre::Result;
 
-pub fn run_local(config: Config) -> AppResult<()> {
+pub fn run_local(config: Config) -> Result<()> {
     commands::run(config)
 }
 
@@ -18,6 +17,6 @@ pub fn run_git(
     branch: Option<&str>,
     commit: Option<&str>,
     config: Config,
-) -> AppResult<()> {
+) -> Result<()> {
     git::run(repo_url, branch, commit, config)
 }